second round review fixes

+++ src/policies/accounting/RecurringAllowance.sol
@@ -15,21 +15,20 @@ library RecurringAllowance {
     struct Limit {
         /// @dev Maximum spend per period window.
         uint160 allowance;
-        /// @dev Period length in seconds. uint48 (not uint40) so that `PeriodUsage` packs into exactly
-        ///      one storage slot: uint48 + uint48 + uint160 = 256 bits.
-        uint48 period;
+        /// @dev Period length in seconds.
+        uint40 period;
         /// @dev Start timestamp (seconds) inclusive.
-        uint48 start;
+        uint40 start;
         /// @dev End timestamp (seconds) exclusive.
-        uint48 end;
+        uint40 end;
     }
 
     /// @notice Stored usage snapshot for a particular active period.
     struct PeriodUsage {
-        /// @dev Period start timestamp (seconds). uint48 so this struct packs into one slot (uint48 + uint48 + uint160 = 256 bits).
-        uint48 start;
+        /// @dev Period start timestamp (seconds).
+        uint40 start;
         /// @dev Period end timestamp (seconds).
-        uint48 end;
+        uint40 end;
         /// @dev Amount spent during the period window.
         uint160 spend;
     }
@@ -51,13 +50,13 @@ library RecurringAllowance {
     error ZeroAllowance();
 
     /// @notice Thrown when `start >= end` in the limit bounds.
-    error InvalidStartEnd(uint48 start, uint48 end);
+    error InvalidStartEnd(uint40 start, uint40 end);
 
     /// @notice Thrown when the current timestamp is before the allowance window.
-    error BeforeStart(uint48 currentTimestamp, uint48 start);
+    error BeforeStart(uint40 currentTimestamp, uint40 start);
 
     /// @notice Thrown when the current timestamp is at or past the allowance window end.
-    error AfterEnd(uint48 currentTimestamp, uint48 end);
+    error AfterEnd(uint40 currentTimestamp, uint40 end);
 
     /// @notice Thrown when the spend value is zero.
     error ZeroValue();
@@ -125,7 +124,7 @@ library RecurringAllowance {
         view
         returns (PeriodUsage memory)
     {
-        uint48 currentTimestamp = uint48(block.timestamp);
+        uint40 currentTimestamp = uint40(block.timestamp);
         if (currentTimestamp < limit.start) revert BeforeStart(currentTimestamp, limit.start);
         if (currentTimestamp >= limit.end) revert AfterEnd(currentTimestamp, limit.end);
 
@@ -134,10 +133,10 @@ library RecurringAllowance {
         bool lastStillActive = currentTimestamp < lastUpdated.end;
         if (lastExists && lastStillActive) return lastUpdated;
 
-        uint48 currentPeriodProgress = (currentTimestamp - limit.start) % limit.period;
-        uint48 start = currentTimestamp - currentPeriodProgress;
+        uint40 currentPeriodProgress = (currentTimestamp - limit.start) % limit.period;
+        uint40 start = currentTimestamp - currentPeriodProgress;
         bool endOverflow = uint256(start) + uint256(limit.period) > limit.end;
-        uint48 end = endOverflow ? limit.end : start + limit.period;
+        uint40 end = endOverflow ? limit.end : start + limit.period;
         return PeriodUsage({start: start, end: end, spend: 0});
     }
 }
